requirements.txt
djangorestframework
psycopg[binary]
python-dotenv
SQLAlchemy
requests
django-cors-headers
